Push WinoBias and StereoSet experiment code and result
Push WinoBias and StereoSet experiment code and result. The commit includes the PROCESSED datasets that can be fed into /scripts/metric.py . Also the code metric.py was modified to support scientific models and the new data format. The .slurm files are the sbatch submission files. The outputs are in the respective slurm outpur folders.